// frontend/src/api/models/handlerCustomerResponse.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { HandlerCustomerResponseLevel } from './handlerCustomerResponseLevel'
import type { HandlerCustomerResponseStatus } from './handlerCustomerResponseStatus'
import type { HandlerCustomerResponseType } from './handlerCustomerResponseType'

/**
 * Customer details returned by the API
 */
export interface HandlerCustomerResponse {
  address?: string
  attributes?: string
  balance?: number
  city?: string
  code?: string
  contact_name?: string
  country?: string
  created_at?: string
  credit_limit?: number
  email?: string
  full_address?: string
  id?: string
  level?: HandlerCustomerResponseLevel
  name?: string
  notes?: string
  phone?: string
  postal_code?: string
  province?: string
  short_name?: string
  sort_order?: number
  status?: HandlerCustomerResponseStatus
  tax_id?: string
  tenant_id?: string
  type?: HandlerCustomerResponseType
  updated_at?: string
  version?: number
}