// frontend/src/api/models/handlerCustomerResponseStatus.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type HandlerCustomerResponseStatus =
  (typeof HandlerCustomerResponseStatus)[keyof typeof HandlerCustomerResponseStatus]

export const HandlerCustomerResponseStatus = {
  active: 'active',
  inactive: 'inactive',
  suspended: 'suspended',
} as const