// frontend/src/api/models/handlerCustomerResponseLevel.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type HandlerCustomerResponseLevel =
  (typeof HandlerCustomerResponseLevel)[keyof typeof HandlerCustomerResponseLevel]

export const HandlerCustomerResponseLevel = {
  normal: 'normal',
  silver: 'silver',
  gold: 'gold',
  platinum: 'platinum',
  vip: 'vip',
} as const